Move createOption to the container
