+++ src/components/DropdownFilter.ts
@@ -1,4 +1,4 @@
-import { Component, OptionHTMLAttributes, ReactElement, createElement } from "react";
+import { Component, ReactElement, createElement } from "react";
 import * as classNames from "classNames";
 
 import "../ui/DropdownFilter.css";
@@ -6,6 +6,7 @@ import "../ui/DropdownFilter.css";
 interface DropdownFilterProps {
     caption: string;
     value: string;
+    getDropdownOptions?: Array<ReactElement<{}>>;
     getFilterValue?: (query: string) => void;
     options: Array<{ filterOptionAttribute: string, filterOptionValue: string}>;
 }
@@ -27,7 +28,7 @@ export class DropdownFilter extends Component<DropdownFilterProps, DropdownFilte
         return (
             createElement("label", { className: classNames("caption-display") }, this.props.caption + ": ",
                 createElement("select", { className: classNames("dropdown-display") }, this.props.options,
-                    this.createOptions))
+                    this.props.getDropdownOptions))
         );
     }
 
@@ -45,21 +46,5 @@ export class DropdownFilter extends Component<DropdownFilterProps, DropdownFilte
 
     private resetQuery() {
         this.setState({ query: "" });
-    }
-
-    private createOptions(props: DropdownFilterProps): Array<ReactElement<{}>> {
-        const optionElements: Array<ReactElement<{}>> = [];
-        if (props.options.length) {
-            props.options.map((optionObject) => {
-                const { filterOptionAttribute, filterOptionValue } = optionObject;
-                const optionValue: OptionHTMLAttributes<HTMLOptionElement> = {
-                    className: "",
-                    label: filterOptionValue,
-                    value: filterOptionAttribute
-                };
-                optionElements.push(createElement("option", optionValue));
-            });
-        }
-        return optionElements;
     }
 }
